예약 페이지 대기 이벤트에서 안 보이도록 수정
