+++ src/main/java/com/multi/hereevent/event/EventController.java
@@ -229,80 +229,5 @@ public class EventController {
 
     }
 
-    /***** 관리자 페이지 *****/
-    @GetMapping("/admin/event")
-    public String selectEventWithPage(@RequestParam Map<String, Object> params,
-                                      @PageableDefault(value = 10,sort = "event_no", direction = Sort.Direction.DESC) Pageable page, Model model){
-        Page<EventDTO> result = eventService.selectEventWithPage(params, page);
-        System.out.println("param==>"+params);
-        System.out.println("page==>"+page);
-        model.addAttribute("type", params.get("type"));
-        model.addAttribute("keyword", params.get("keyword"));
-        model.addAttribute("eventList", result.getContent());
-        model.addAttribute("totalPages", result.getTotalPages());
-        model.addAttribute("totalElements", result.getTotalElements());
-        model.addAttribute("pageNumber", page.getPageNumber());
-        return "admin/event";
-    }
-    // insert, update, delete 만들어 놨는데 수정해서 쓰시면 될거같습니다.
-    @GetMapping("/admin/event/insert")
-    public String createEventPage(Model model){
-        List<CategoryDTO> categoryList = new ArrayList<>();
-        categoryList = categoryService.getListCategory();
-        model.addAttribute("categoryList",categoryList);
-        return "event/insert";
-    }
-    @PostMapping("/admin/event/insert")
-    public String createEvent(EventDTO event) {
-        System.out.println(event);
-        //주소 합쳐서 넣기
-        event.setAddr(event.getAddr()+event.getDetailAddress()+event.getExtraAddress());
-        //행사 이미지 등록하기
-        MultipartFile eventImg = event.getEvent_img();
-        String storeFilename = null;
-        try {
-            storeFilename = fileUploadService.uploadEventImg(eventImg);
-            event.setImg_path(storeFilename);
-            //DB에 삽입
-            eventService.insertEvent(event);
-            System.out.println("+++++"+event);
-            return "redirect:/admin/event";
-        } catch (IOException e) {
-            new RuntimeException();
-            return "common/errorPage";
-        }
-    }
-    @GetMapping("/admin/event/update/{event_no}")
-    public String updateEventPage(@PathVariable("event_no") int event_no, Model model){
-        List<CategoryDTO> categoryList = new ArrayList<>();
-        categoryList = categoryService.getListCategory();
-        model.addAttribute("categoryList",categoryList);
-        EventDTO event = eventService.getEventDetails(event_no);
-        System.out.println(event);
-        model.addAttribute("event",event);
-        return "event/update";
-    }
-    @PostMapping("/admin/event/update")
-    public String updateEvent(@RequestParam("event_no") int event_no,EventDTO event) {
-        System.out.println("updateEvent====>"+event);
-        MultipartFile eventImg = event.getEvent_img();
-        String storeFilename = null;
-        try {
-            storeFilename = fileUploadService.uploadEventImg(eventImg);
-            event.setImg_path(storeFilename);
-            event.setAddr(event.getAddr()+event.getDetailAddress()+event.getExtraAddress());
-            eventService.updateEvent(event);
-            return "redirect:/admin/event";
-        } catch (IOException e) {
-            new RuntimeException(e);
-            return "common/errorPage";
-        }
-    }
 
-    @PostMapping("/admin/event/delete")
-    public String deleteEvent(@RequestParam("eventNo") List<Integer> eventNo) {
-        System.out.println(eventNo);
-        eventService.deleteEvent(eventNo);
-        return "redirect:/admin/event";
-    }
 }
