+++ src/main/resources/templates/detailedPage/detailedPage.html
@@ -135,7 +135,7 @@
                         <div class="submenu">
                             <a href="javascript:void(0);" class="on" onclick="scrollToSection('content-section');activateTab(event)" th:classappend="${tab == 'content'} ? 'active' : ''" data-tab="content">상세정보</a>
                             <a href="javascript:void(0);" onclick="scrollToSection('navigation-section');activateTab(event)" th:classappend="${tab == 'map'} ? 'active' : ''" data-tab="map">길찾기</a>
-                            <a href="javascript:void(0);" onclick="scrollToSection('reservation-section');activateTab(event)" th:classappend="${tab == 'reservation'} ? 'active' : ''" data-tab="reservation">예약</a>
+                            <a href="javascript:void(0);" onclick="scrollToSection('reservation-section');activateTab(event)" th:classappend="${tab == 'reservation'} ? 'active' : ''" th:unless="${event.type != 'wait'}" data-tab="reservation">예약</a>
                             <a href="javascript:void(0);" onclick="scrollToSection('reviews-section');activateTab(event)" th:classappend="${tab == 'reviews'} ? 'active' : ''" data-tab="reviews">후기</a>
                         </div>
 
@@ -149,9 +149,9 @@
                             <th:block th:replace="~{detailedPage/navigation :: navigation}"></th:block>
                         </article>
 
-                        <article id="reservation-section" class="article">
+                        <article id="reservation-section" class="article" th:unless="${event.type != 'wait'}">
                             <!-- 예약 내용 -->
-                            <th:block th:replace="~{detailedPage/reservation :: reservation}"></th:block>
+                            <th:block th:replace="~{detailedPage/reservation :: reservation}" ></th:block>
                         </article>
 
                         <article id="reviews-section" class="article">
